fix(util): move uuid to constants
