+++ nuxt/pages/task/account/email-address/verify.vue
@@ -9,7 +9,7 @@
 import consola from 'consola'
 import Swal from 'sweetalert2'
 
-import { REGEX_UUID } from '~/plugins/util/validation'
+import { REGEX_UUID } from '~/plugins/util/constants'
 import { getApiMeta } from '~/plugins/util/util'
 import { useAccountEmailAddressVerificationMutation } from '~/gql/generated'
